dev: renamed app error module, ErrDb -> ErrDB
